+++ lib/td_dd/grants/requests.ex
@@ -9,7 +9,6 @@ defmodule TdDd.Grants.Requests do
   alias Ecto.Multi
   alias TdCache.Permissions
   alias TdCache.UserCache
-  alias TdDd.Auth.Claims
   alias TdDd.DataStructures
   alias TdDd.DataStructures.DataStructure
   alias TdDd.Grants.Audit
@@ -18,6 +17,9 @@ defmodule TdDd.Grants.Requests do
   alias TdDd.Grants.GrantRequestGroup
   alias TdDd.Grants.GrantRequestStatus
   alias TdDd.Repo
+  alias Truedat.Auth.Claims
+
+  defdelegate authorize(action, user, params), to: TdDd.Grants.Policy
 
   def list_grant_request_groups do
     Repo.all(GrantRequestGroup)
@@ -221,9 +223,10 @@ defmodule TdDd.Grants.Requests do
 
   def latest_grant_request_by_data_structure(data_structure_id, user_id) do
     GrantRequest
-    |> where([gr], data_structure_id: ^data_structure_id)
-    |> join(:left, [gr], grg in assoc(gr, :group))
-    |> where([_, gr], gr.user_id == ^user_id)
+    |> where([r], data_structure_id: ^data_structure_id)
+    |> join(:inner, [r], g in assoc(r, :group))
+    |> where([_, g], g.user_id == ^user_id)
+    |> preload(:group)
     |> order_by(desc: :inserted_at)
     |> limit(1)
     |> Repo.one()
@@ -421,7 +424,6 @@ defmodule TdDd.Grants.Requests do
     %{grant_request | pending_roles: pending_roles}
   end
 
-  # @spec with_missing_roles(map, MapSet.t(), map())
   defp with_missing_roles(
          %{approvals: approvals, domain_ids: domain_ids} = grant_request,
          required_roles,