+++ lib/td_dd_web/controllers/data_structure_controller.ex
@@ -2,8 +2,7 @@ defmodule TdDdWeb.DataStructureController do
   use TdDdWeb, :controller
   use PhoenixSwagger
 
-  import Canada, only: [can?: 2]
-  import Canada.Can, only: [can?: 3]
+  import Bodyguard, only: [permit?: 4]
 
   alias TdDd.CSV.Download
   alias TdDd.DataStructures
@@ -13,7 +12,7 @@ defmodule TdDdWeb.DataStructureController do
   alias TdDd.DataStructures.DataStructure
   alias TdDd.DataStructures.DataStructureVersion
   alias TdDd.DataStructures.Search
-  alias TdDd.DataStructures.StructureNote
+  alias TdDd.DataStructures.StructureNotes
   alias TdDd.DataStructures.StructureNotesWorkflow
   alias TdDd.DataStructures.Tags
   alias TdDd.Utils.FileHash
@@ -37,8 +36,6 @@ defmodule TdDdWeb.DataStructureController do
     :metadata_versions,
     :data_structure_type
   ]
-
-  @structures_actions [:update_domain_ids]
 
   plug(TdDdWeb.SearchPermissionPlug)
 
@@ -99,7 +96,7 @@ defmodule TdDdWeb.DataStructureController do
 
     with %{valid?: true} = changeset <-
            DataStructures.update_changeset(claims, structure, structure_params),
-         {:can, true} <- {:can, can?(claims, update_data_structure(changeset))},
+         :ok <- Bodyguard.permit(DataStructures, :update_data_structure, claims, changeset),
          {:ok, _} <- DataStructures.update_data_structure(claims, changeset, inherit) do
       claims = conn.assigns[:current_resource]
       options = filter_opts(claims, structure)
@@ -114,7 +111,7 @@ defmodule TdDdWeb.DataStructureController do
   defp filter_opts(claims, structure) do
     Enum.filter(@enrich_attrs, fn
       :with_protected_metadata ->
-        can?(claims, view_protected_metadata(structure))
+        Bodyguard.permit?(DataStructures, :view_protected_metadata, claims, structure)
 
       _ ->
         true
@@ -139,16 +136,10 @@ defmodule TdDdWeb.DataStructureController do
   def delete(conn, %{"id" => id, "logical" => "true"}) do
     claims = conn.assigns[:current_resource]
 
-    with %DataStructure{} = data_structure <-
-           DataStructures.get_data_structure!(id),
-         {:can, true} <- {:can, can?(claims, delete_structure(data_structure))},
-         %DataStructureVersion{} = data_structure_version <-
-           DataStructures.get_latest_version(data_structure),
-         {:ok, _} <-
-           DataStructures.logical_delete_data_structure(
-             data_structure_version,
-             claims
-           ) do
+    with %DataStructure{} = ds <- DataStructures.get_data_structure!(id),
+         :ok <- Bodyguard.permit(DataStructures, :delete_data_structure, claims, ds),
+         %DataStructureVersion{} = dsv <- DataStructures.get_latest_version(ds),
+         {:ok, _} <- DataStructures.logical_delete_data_structure(dsv, claims) do
       send_resp(conn, :no_content, "")
     end
   end
@@ -157,7 +148,7 @@ defmodule TdDdWeb.DataStructureController do
     claims = conn.assigns[:current_resource]
     data_structure = DataStructures.get_data_structure!(id)
 
-    with {:can, true} <- {:can, can?(claims, delete_data_structure(data_structure))},
+    with :ok <- Bodyguard.permit(DataStructures, :delete_data_structure, claims, data_structure),
          {:ok, %{data_structure: _deleted_data_structure}} <-
            DataStructures.delete_data_structure(data_structure, claims) do
       send_resp(conn, :no_content, "")
@@ -217,7 +208,7 @@ defmodule TdDdWeb.DataStructureController do
     permission = conn.assigns[:search_permission]
     auto_publish = bulk_update_request |> Map.get("auto_publish", false)
 
-    with {:can, true} <- {:can, can?(claims, create(BulkUpdate))},
+    with :ok <- Bodyguard.permit(BulkUpdate, :bulk_update_notes, claims),
          %{results: results} <- search_all_structures(claims, permission, search_params),
          ids <- Enum.map(results, & &1.id),
          {:ok, %{update_notes: update_notes}} <-
@@ -265,9 +256,8 @@ defmodule TdDdWeb.DataStructureController do
     headers = ["external_id", "domain_external_ids"]
 
     with claims <- conn.assigns[:current_resource],
-         {:can, true} <- {:can, can?(claims, update_domain_ids(DataStructure))},
-         :ok <-
-           BulkUpdate.check_csv_headers(structures_content_upload, headers),
+         :ok <- Bodyguard.permit(BulkUpdate, :bulk_upload_domains, claims),
+         :ok <- BulkUpdate.check_csv_headers(structures_content_upload, headers),
          [_ | _] = rows <- BulkUpdate.from_csv(structures_content_upload, :simple),
          info <- BulkUpdate.csv_bulk_update_domains(rows, claims),
          summary <- csv_bulk_upload_domains_summary(rows, info) do
@@ -379,17 +369,16 @@ defmodule TdDdWeb.DataStructureController do
         true
 
       {_content, %{data_structure: data_structure}} ->
-        action = StructureNotesWorkflow.get_action_editable_action(data_structure)
-
         can_edit =
-          case action do
-            :create -> can?(claims, create_structure_note(data_structure))
-            :edit -> can?(claims, edit_structure_note(data_structure))
+          case StructureNotesWorkflow.get_action_editable_action(data_structure) do
+            :create -> permit?(StructureNotes, :create, claims, data_structure)
+            :edit -> permit?(StructureNotes, :edit, claims, data_structure)
             _ -> true
           end
 
         if auto_publish do
-          can_edit and can?(claims, publish_structure_note_from_draft(data_structure))
+          can_edit and
+            permit?(StructureNotes, :publish_draft, claims, data_structure)
         else
           can_edit
         end
@@ -482,11 +471,13 @@ defmodule TdDdWeb.DataStructureController do
   end
 
   defp do_render_data_structure(conn, claims, data_structure) do
-    if can?(claims, view_data_structure(data_structure)) do
+    if permit?(DataStructures, :view_data_structure, claims, data_structure) do
       user_permissions = %{
-        update: can?(claims, update_data_structure(data_structure)),
-        confidential: can?(claims, manage_confidential_structures(data_structure)),
-        view_profiling_permission: can?(claims, view_data_structures_profile(data_structure))
+        update: permit?(DataStructures, :update_data_structure, claims, data_structure),
+        confidential:
+          permit?(DataStructures, :manage_confidential_structures, claims, data_structure),
+        view_profiling_permission:
+          permit?(DataStructures, :view_data_structures_profile, claims, data_structure)
       }
 
       # TODO: tags not consumed by front?
@@ -551,8 +542,8 @@ defmodule TdDdWeb.DataStructureController do
 
     actions =
       params
-      |> actions_notes(total)
-      |> Enum.filter(&can?(claims, &1, StructureNote))
+      |> bulk_actions(total)
+      |> Enum.filter(&Bodyguard.permit?(BulkUpdate, &1, claims))
       |> Enum.reduce(%{}, fn
         :bulk_update, acc ->
           Map.put(acc, "bulkUpdate", %{
@@ -571,28 +562,20 @@ defmodule TdDdWeb.DataStructureController do
             href: Routes.data_structure_path(conn, :bulk_update_template_content),
             method: "POST"
           })
+
+        :bulk_upload_domains, acc ->
+          Map.put(acc, "bulkUploadDomains", %{
+            href: Routes.data_structure_path(conn, :bulk_upload_domains),
+            method: "POST"
+          })
       end)
-      |> put_actions_structures(conn, claims)
 
     assign(conn, :actions, actions)
   end
 
-  defp put_actions_structures(actions, conn, claims) do
-    @structures_actions
-    |> Enum.filter(&can?(claims, &1, DataStructure))
-    |> Enum.reduce(%{}, fn
-      :update_domain_ids, acc ->
-        Map.put(acc, "bulkUploadDomains", %{
-          href: Routes.data_structure_path(conn, :bulk_upload_domains),
-          method: "POST"
-        })
-    end)
-    |> Map.merge(actions)
+  defp bulk_actions(%{"filters" => %{"type.raw" => [_]}} = _params, total) when total > 0 do
+    [:bulk_upload, :auto_publish, :bulk_update, :bulk_upload_domains]
   end
 
-  defp actions_notes(%{"filters" => %{"type.raw" => [_]}} = _params, total) when total > 0 do
-    [:bulk_upload, :auto_publish, :bulk_update]
-  end
-
-  defp actions_notes(_params, _), do: [:bulk_upload, :auto_publish]
+  defp bulk_actions(_params, _), do: [:bulk_upload, :auto_publish, :bulk_upload_domains]
 end