+++ lib/td_dd_web/controllers/grant_request_status_controller.ex
@@ -1,8 +1,6 @@
 defmodule TdDdWeb.GrantRequestStatusController do
   use TdDdWeb, :controller
 
-  import Canada, only: [can?: 2]
-
   alias TdDd.Grants.Requests
   alias TdDd.Grants.Statuses
   alias TdDdWeb.GrantRequestView
@@ -12,7 +10,7 @@ defmodule TdDdWeb.GrantRequestStatusController do
   def create(conn, %{"grant_request_id" => id, "status" => "cancelled" = status}) do
     with %{user_id: user_id} = claims <- conn.assigns[:current_resource],
          request <- Requests.get_grant_request!(id, claims),
-         {:can, true} <- {:can, can?(claims, cancel(request))},
+         :ok <- Bodyguard.permit(Requests, :cancel, claims, request),
          {:ok, _grant_request_status} <-
            Statuses.create_grant_request_status(request, status, user_id),
          updated_request <- Requests.get_grant_request!(id, claims) do
@@ -28,7 +26,7 @@ defmodule TdDdWeb.GrantRequestStatusController do
 
     with %{user_id: user_id} = claims <- conn.assigns[:current_resource],
          request <- Requests.get_grant_request!(id, claims),
-         {:can, true} <- {:can, can?(claims, approve(request))},
+         :ok <- Bodyguard.permit(Requests, :approve, claims, request),
          {:ok, _grant_request_status} <-
            Statuses.create_grant_request_status(request, status, user_id, status_reason),
          updated_request <- Requests.get_grant_request!(id, claims) do