+++ lib/td_dq/auth/pipeline/unsecure.ex
@@ -3,7 +3,7 @@ defmodule TdDq.Auth.Pipeline.Unsecure do
   use Guardian.Plug.Pipeline,
     otp_app: :td_dd,
     error_handler: Truedat.Auth.ErrorHandler,
-    module: TdDq.Auth.Guardian
+    module: Truedat.Auth.Guardian
 
   plug Guardian.Plug.VerifyHeader
   plug Guardian.Plug.LoadResource, allow_blank: true